[BattlePets]: DB support for previous commit / fix compile

+++ src/server/game/BattlePet/BattlePetMgr.h
@@ -83,7 +83,7 @@ public:
     void FillBattlePetJournal();
     void BuildBattlePetJournal(WorldPacket *data);
 
-    void AddBattlePetInJournal(uint64 guid, uint32 speciesID, uint32 creatureEntry, uint32 level, uint32 display, uint32 power, uint32 speed, uint32 health, uint32 maxHealth, uint32 quality, uint32 xp, uint32 flags, uint32 spellID, std::string customName = "", int16 breedID = -1);
+    void AddBattlePetInJournal(uint64 guid, uint32 speciesID, uint32 creatureEntry, uint8 level, uint32 display, uint16 power, uint16 speed, uint32 health, uint32 maxHealth, uint8 quality, uint16 xp, uint16 flags, uint32 spellID, std::string customName = "", int16 breedID = -1);
 
     void SendClosePetBattle();
     void SendUpdatePets(uint8 petCount);
